Separate the shared GPU-sharing machinery from time-slicing
Pure refactor: no behaviour change, and time-slicing is still the only caller.
Splitting it out first so the MPS feature that follows reads as a feature rather than
a feature plus a rearrangement of everything around it.

Time-slicing had accumulated the generic parts of GPU sharing inside the file named
for it -- the device-plugin swap, the deletion and drain helpers, the rollout watcher
and failure dumps, concurrency observation, log and marker parsing. A second feature
that shares a GPU needs all of it, so a file named time_slicing_test.go owning it
would make the next feature depend on symbols nothing indicates are shared.

gpu_sharing_test.go holds the neutral parts, gpu_diagnostics_test.go the rollout
watcher and dumps, and time_slicing_test.go drops from 735 lines to 271 of
time-slicing.

Three things change shape rather than just moving:

The diagnostics take a sharingFeature -- a name for the log prefix and the ConfigMap
the plugin mounts -- instead of hardcoding time-slicing's. They currently only ever
receive time-slicing's, so this is inert here, but hardcoding it is exactly how a
second caller ends up reporting the wrong feature's name and inspecting the wrong
ConfigMap when it fails.

Per-invocation state moves from package-level variables into a sharingState the
feature closures capture. The flags recording "the plugin has been swapped" and "the
replacement came up" belong to one run; as globals a second invocation in the same
binary would inherit the previous run's values.

The plugin image comes from the stock manifest at render time rather than being
repeated in the reconfigured one, so a version bump in test/manifests/assets cannot
leave this copy behind.

job-time-slicing-workload.yaml becomes job-gpu-sharing-workload.yaml and templates
its Job name, since nothing in it is specific to time-slicing.

Verified by running the time-slicing feature on g5.8xlarge after the change:
allocatable 1 -> 10, ten pods concurrent on one A10G, same GPU UUID, stock plugin
restored.
