+++ test/cases/nvidia/manifests/time-slicing-device-plugin.yaml
@@ -52,7 +52,7 @@ spec:
         effect: NoSchedule
       priorityClassName: "system-node-critical"
       containers:
-      - image: nvcr.io/nvidia/k8s-device-plugin:v0.17.2
+      - image: {{.PluginImage}}
         name: nvidia-device-plugin-ctr
         env:
           - name: FAIL_ON_INIT_ERROR
